// src/types.ts
export interface PropertyData {
  listed_by: string;
  property_status: string;
  property_ref: string;
  bedrooms: string;
  bathrooms: string;
  created_at: string;
  full_count: string;
  id: string;
  url: string;
  images: string;
  originalimg: string;
  short_address: string;
  formatted_address: string;
  price: number;
  available_from: string;
  asking_price: number;
  marker_latitude: string;
  marker_longitude: string;
  description: string;
  position: string;
  updated_at: string;
  rn: string;
  overall_count: string;
}